@@ -1,8 +1,9 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './components/App'
+import * as ReactDOMClient from 'react-dom/client';
+import App from './components/App';
 
-ReactDOM.render(
-  <App />,
-  document.getElementById('root')
+const root = ReactDOMClient.createRoot(document.getElementById('root'));
+
+root.render(
+  <App />
 );
